Init: Initial commit of this app
This commit, in its current form, is a stable release of a basic
AngularJS v1.6.6 application with basic CRUD functions. The application
does have room for improvement, but it currently is an acceptable
demonstration of the MV* architecture of a basic AngularJS application.

TODO: Improve how data is displayed in the edit view of the app so that
previous values are provided as default values rather than as
placeholders.